requirements.txt
addict==2.4.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
e3nn==0.5.1
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
fire==0.6.0
Flask==3.0.3
fsspec==2024.6.0
importlib_metadata==7.1.0
iopath==0.1.9
ipython==8.18.1
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
Mako==1.3.5
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
open3d==0.18.0
opencv-python==4.10.0.82
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
plotly==5.22.0
ply==3.11
portalocker==1.4.0
prompt_toolkit==3.0.46
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyOpenGL==3.1.7
PyPatchMatch==1.0.1
PyQt5==5.15.10
pyquaternion==0.9.9
pytools==2024.1.5
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@f34104cf6ebefacd7b7e07955ee7aaa823e616ac
pytz==2024.1
referencing==0.35.1
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.0
simple-lama-inpainting==0.1.0
stack-data==0.6.3
tenacity==8.3.0
threadpoolctl==3.5.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
traitlets==5.14.3
triton==2.2.0
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.11
